+++ core/src/main/kotlin/no/elg/hex/screens/LevelSelectScreen.kt
@@ -117,6 +117,13 @@ class LevelSelectScreen : AbstractScreen(), ReloadableScreen {
             font.draw(batch, layout, x, y + vertOffset)
           }
 
+          if (metadata.userRoundsToBeat != Island.NEVER_PLAYED) {
+            val color = if (metadata.isUserBetterThanAuthor()) Color.GOLD else Color.WHITE
+            val align = if(showIslandId) Align.right else Align.center
+            layout.setText(font, "Best: ${metadata.userRoundsToBeat} rounds", color, width, align, true)
+            font.draw(batch, layout, x, y + vertOffset)
+          }
+
           if (Hex.mapEditor || Hex.debug) {
             layout.setText(font, "ARtB ${metadata.authorRoundsToBeat}", Color.WHITE, width, Align.left, true)
             font.draw(batch, layout, x, y + vertOffset * 2)