+++ CHANGES.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add button to create new islands directly from the level creation screen
 * New islands validation rule: There must be at least two different teams with a capital
 * Show the island id to make it easier to communicate about islands
+* Show high score for each island in the level select screen
 
 ### Changed